Efficient snake
This is a complete overhaul on how the snake works. The entirety of the snake and gridpane are no longer used to calculate the path of the snake. Only the head and tail will update.

No more "DARKCYAN" the entire gridpane just to move one square.

The snake is updated by adding to the begining of the arraylist and dropping the end insted of using a for loop to assign (n-1)'s to n.

This decreases calculations from O(n^2) to O(n).

There may be other things that I need to find that are O(n^2).

Also Dany, I left some notes on how to fix some problems I found in your AI.